Код из main.js вынесен в отдельные модули
